+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0),
 
 ## Unreleased
 
-Nothing.
+### Security
+
+- Added a timeout of 3 seconds to HTTP server
+  ([GO-S2114](https://deepsource.io/directory/analyzers/go/issues/GO-S2114),
+  [CWE-400](https://cwe.mitre.org/data/definitions/400.html)).
 
 ## [1.0.0](https://github.com/trallnag/token2go-server/compare/3d62e4caf205bdf26b12b3900e27540e6ebfbd2e...v1.0.0) / 2022-12-17